build: place NAPI-related changes behind a configure flag
Fixes https://github.com/nodejs/abi-stable-node/issues/80
Closes https://github.com/nodejs/abi-stable-node/pull/89
