+++ Makefile
@@ -9,6 +9,7 @@ FLAKY_TESTS ?= run
 TEST_CI_ARGS ?=
 STAGINGSERVER ?= node-www
 LOGLEVEL ?= silent
+NAPI ?= False
 OSTYPE := $(shell uname -s | tr '[A-Z]' '[a-z]')
 
 ifdef JOBS
@@ -120,12 +121,17 @@ v8:
 	tools/make-v8.sh
 	$(MAKE) -C deps/v8 $(V8_ARCH).$(BUILDTYPE_LOWER) $(V8_BUILD_OPTIONS)
 
+ifeq ($(NAPI),True)
+BUILD_ADDONS = build-addons build-addons-abi
+TESTS_TO_RUN = addons addons-abi doctool inspector known_issues message pseudo-tty parallel sequential
+else
+BUILD_ADDONS = build-addons
+TESTS_TO_RUN = addons doctool inspector known_issues message pseudo-tty parallel sequential
+endif
+
 test: all
-	$(MAKE) build-addons
-	$(MAKE) build-addons-abi
-	$(MAKE) cctest
-	$(PYTHON) tools/test.py --mode=release -J \
-		addons addons-abi doctool inspector known_issues message pseudo-tty parallel sequential
+	$(MAKE) $(BUILD_ADDONS) cctest
+	$(PYTHON) tools/test.py --mode=release -J $(TESTS_TO_RUN)
 	$(MAKE) lint
 
 test-parallel: all
@@ -189,6 +195,8 @@ test/addons/.buildstamp: config.gypi \
 # TODO(bnoordhuis) Force rebuild after gyp update.
 build-addons: $(NODE_EXE) test/addons/.buildstamp
 
+ifeq ($(NAPI),True)
+
 ADDONS_ABI_BINDING_GYPS := \
 	$(filter-out test/addons-abi/??_*/binding.gyp, \
 		$(wildcard test/addons-abi/*/binding.gyp))
@@ -215,15 +223,18 @@ test/addons-abi/.buildstamp: $(ADDONS_ABI_BINDING_GYPS) \
 # TODO(bnoordhuis) Force rebuild after gyp or node-gyp update.
 build-addons-abi: $(NODE_EXE) test/addons-abi/.buildstamp
 
+endif #eq ($(NAPI),True)
 
 test-gc: all test/gc/node_modules/weak/build/Release/weakref.node
 	$(PYTHON) tools/test.py --mode=release gc
 
-test-build: | all build-addons build-addons-abi
+test-build: | all $(BUILD_ADDONS)
 
 test-build-addons: all build-addons
 
+ifeq ($(NAPI),True)
 test-build-addons-abi: all build-addons-abi
+endif #eq ($(NAPI),True)
 
 test-all: test-build test/gc/node_modules/weak/build/Release/weakref.node
 	$(PYTHON) tools/test.py --mode=debug,release
@@ -231,7 +242,7 @@ test-all: test-build test/gc/node_modules/weak/build/Release/weakref.node
 test-all-valgrind: test-build
 	$(PYTHON) tools/test.py --mode=debug,release --valgrind
 
-CI_NATIVE_SUITES := addons addons-abi
+CI_NATIVE_SUITES := addons $(if $(filter $(NAPI),True),addons-abi)
 CI_JS_SUITES := doctool inspector known_issues message parallel pseudo-tty sequential
 
 # Build and test addons without building anything else
@@ -248,7 +259,7 @@ test-ci-js:
 		$(TEST_CI_ARGS) $(CI_JS_SUITES)
 
 test-ci: LOGLEVEL := info
-test-ci: | build-addons build-addons-abi
+test-ci: | $(BUILD_ADDONS)
 	$(PYTHON) tools/test.py $(PARALLEL_ARGS) -p tap --logfile test.tap \
 		--mode=release --flaky-tests=$(FLAKY_TESTS) \
 		$(TEST_CI_ARGS) $(CI_NATIVE_SUITES) addon-abi $(CI_JS_SUITES)
@@ -292,8 +303,10 @@ test-npm-publish: $(NODE_EXE)
 test-addons: test-build-addons
 	$(PYTHON) tools/test.py --mode=release addons
 
+ifeq ($(NAPI),True)
 test-addons-abi: test-build-addons-abi
 	$(PYTHON) tools/test.py --mode=release addons-abi
+endif #eq ($(NAPI),True)
 
 test-timers:
 	$(MAKE) --directory=tools faketime
@@ -753,7 +766,9 @@ CPPLINT_EXCLUDE += src/node_root_certs.h
 CPPLINT_EXCLUDE += src/queue.h
 CPPLINT_EXCLUDE += src/tree.h
 CPPLINT_EXCLUDE += $(wildcard test/addons/??_*/*.cc test/addons/??_*/*.h)
+ifeq ($(NAPI),True)
 CPPLINT_EXCLUDE += $(wildcard test/addons-abi/??_*/*.cc test/addons-abi/??_*/*.h)
+endif #eq ($(NAPI),True)
 
 CPPLINT_FILES = $(filter-out $(CPPLINT_EXCLUDE), $(wildcard \
 	src/*.c \
@@ -763,8 +778,8 @@ CPPLINT_FILES = $(filter-out $(CPPLINT_EXCLUDE), $(wildcard \
 	test/addons/*/*.h \
 	test/cctest/*.cc \
 	test/cctest/*.h \
-	test/addons-abi/*/*.cc \
-	test/addons-abi/*/*.h \
+	$(if $(filter $(NAPI),True),test/addons-abi/*/*.cc) \
+	$(if $(filter $(NAPI),True),test/addons-abi/*/*.h) \
 	tools/icu/*.cc \
 	tools/icu/*.h \
 	))
@@ -794,13 +809,17 @@ lint:
 lint-ci: lint
 endif
 
+ifeq ($(NAPI),True)
+PHONY_ABI = test-addons-abi build-addons-abi
+else
+PHONY_ABI =
+endif #eq ($(NAPI),True)
+
 .PHONY: lint cpplint jslint bench clean docopen docclean doc dist distclean \
 	check uninstall install install-includes install-bin all staticlib \
-	dynamiclib test test-all \
-	test-addons build-addons test-addons-abi build-addons-abi \
-	test-addons-abi	build-addons-abi website-upload pkg \
+	dynamiclib test test-all test-addons build-addons website-upload pkg \
 	blog blogclean tar binary release-only bench-http-simple bench-idle \
 	bench-all bench bench-misc bench-array bench-buffer bench-net \
 	bench-http bench-fs bench-tls cctest run-ci test-v8 test-v8-intl \
 	test-v8-benchmarks test-v8-all v8 lint-ci bench-ci jslint-ci doc-only \
-	$(TARBALL)-headers test-ci test-ci-native test-ci-js build-ci
+	$(TARBALL)-headers test-ci test-ci-native test-ci-js build-ci $(PHONY_ABI)
