+++ node.gyp
@@ -156,7 +156,6 @@
         'src/handle_wrap.cc',
         'src/js_stream.cc',
         'src/node.cc',
-        'src/node_asyncapi.cc',
         'src/node_buffer.cc',
         'src/node_config.cc',
         'src/node_constants.cc',
@@ -165,7 +164,6 @@
         'src/node_file.cc',
         'src/node_http_parser.cc',
         'src/node_javascript.cc',
-        'src/node_jsvmapi.cc',
         'src/node_main.cc',
         'src/node_os.cc',
         'src/node_revert.cc',
@@ -207,10 +205,6 @@
         'src/handle_wrap.h',
         'src/js_stream.h',
         'src/node.h',
-        'src/node_api_helpers.h',
-        'src/node_asyncapi.h',
-        'src/node_asyncapi_internal.h',
-        'src/node_asyncapi_types.h',
         'src/node_buffer.h',
         'src/node_constants.h',
         'src/node_debug_options.h',
@@ -219,9 +213,6 @@
         'src/node_internals.h',
         'src/node_javascript.h',
         'src/node_mutex.h',
-        'src/node_jsvmapi.h',
-        'src/node_jsvmapi_internal.h',
-        'src/node_jsvmapi_types.h',
         'src/node_root_certs.h',
         'src/node_version.h',
         'src/node_watchdog.h',
@@ -265,6 +256,22 @@
 
 
       'conditions': [
+        [ 'enable_napi=="true"', {
+          'sources': [
+            'src/node_asyncapi.cc',
+            'src/node_jsvmapi.cc',
+            'src/node_api_helpers.h',
+            'src/node_asyncapi.h',
+            'src/node_asyncapi_internal.h',
+            'src/node_asyncapi_types.h',
+            'src/node_jsvmapi.h',
+            'src/node_jsvmapi_internal.h',
+            'src/node_jsvmapi_types.h'
+          ],
+          'defines': [
+            'ENABLE_NAPI'
+          ]
+        }],
         [ 'node_shared=="false"', {
           'msvs_settings': {
             'VCManifestTool': {
